requirements.txt
pytest
pytest-watch
pyorbital
requests
apscheduler
Flask
ratelimit
pydispatcher